+++ 2026/Amper/HttpClient/shared/src@mingw/World.kt
@@ -1,0 +1,1 @@
+actual fun getWorld() = "Windows World"